+++ R/gg_stars.R
@@ -22,7 +22,6 @@
 #' @export
 #' @examples
 #' library(simplevis)
-#' library(stars)
 #' 
 #' gg_stars(example_stars, 
 #'      borders = example_borders)
@@ -164,12 +163,9 @@ gg_stars <- function(data,
 #' @export
 #' @examples
 #' library(simplevis)
-#' library(stars)
 #' 
 #' gg_stars_col(example_stars,
 #'              col_var = nitrate,
-#'              col_method = "quantile",
-#'              col_cuts = c(0, 0.05, 0.25, 0.5, 0.75, 0.95, 1),
 #'              col_na_rm = TRUE,
 #'              borders = example_borders)
 #'