+++ R/imports.R
@@ -3,6 +3,7 @@
 #' @import ggplot2
 #' @import leaflet
 #' @import sf
+#' @import stars
 #' @importFrom rlang !!
 #' @importFrom rlang :=
 #' @importFrom rlang .data