+++ inst/shiny/2/make_app_vis.R
@@ -61,5 +61,6 @@ if(map_filter == "None") {
 
 title <- paste0("Monitored trends, 2008\u201317")
 
-leaflet_sf_col(map_data, trend_category, col_title = title)
-
+leaf_sf_col(map_data,
+            col_var = trend_category,
+            col_title = title)